requirements.txt
openai
gradio
sqlalchemy
pymysql
python-dotenv
mysql-connector-python